refactor: centralize tab retrieval logic in getCurrentTab utility
This commit introduces a new utility function, getCurrentTab, to streamline the process of retrieving the current active tab across the CustomSnoozeView, MainView, and RecurringSnoozeView components. The previous inline tab retrieval logic has been replaced with calls to this utility, improving code readability and maintainability. Additionally, the function handles both tab ID query parameters and fallback to the last focused window, enhancing the user experience.
